requirements.txt
pandas
requests
pymongo
azure-storage-blob
azure-functions
pyodbc
sqlalchemy
python-dotenv
psycopg2-binary